さらにリファクタ
