+++ Runtime/Util/Async/TextureLoader.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.Networking;
 
@@ -16,7 +17,7 @@ namespace PLATEAU.Util.Async
         /// </summary>
         public static async Task<Texture2D> LoadAsync(string url, int timeOutSec)
         {
-            System.Uri uri = new System.Uri(url); 
+            Uri uri = new Uri(url); 
             var request = UnityWebRequestTexture.GetTexture(uri);
 
             request.timeout = timeOutSec;
@@ -33,6 +34,65 @@ namespace PLATEAU.Util.Async
 
             var texture = ((DownloadHandlerTexture)request.downloadHandler).texture;
             return texture;
+        }
+
+        
+        /// <summary>
+        /// テクスチャのURL（パス） から、PLATEAUのテクスチャを非同期でロードします。
+        /// 内容は<see cref="LoadAsync"/>に、PLATEAUテクスチャ用の処理を加えたものになります。
+        /// メインスレッドで実行する必要があります。 
+        /// </summary>
+        public static async Task<Texture2D> LoadPlateauTextureAsync(string texturePath)
+        {
+            if (string.IsNullOrEmpty(texturePath))
+                return null;
+
+            // .PLATEAU からの相対パスを求めます。
+            string pathToReplace = (PathUtil.PLATEAUSrcFetchDir + "/").Replace('\\', '/');
+            string relativePath = (texturePath.Replace('\\', '/')).Replace(pathToReplace, "");
+
+            Debug.Log($"Loading Texture : {texturePath}");
+
+            // 非同期でテクスチャをロードします。
+            var texture = await LoadAsync(texturePath, 3);
+
+            if (texture == null)
+                return null;
+
+            // この Compress によってテクスチャ容量が 6分の1 になります。
+            // 松山市のLOD2の建物モデルで計測したところ、 テクスチャのメモリ使用量が 2.6GB から 421.3MB になりました。
+            // 画質は下がりますが、メモリ使用量を適正にするために必須と思われます。
+            var compressedTex = Compress(texture);
+
+            compressedTex.name = relativePath;
+            return compressedTex;
+        }
+        
+        // テクスチャを圧縮します。
+        private static Texture2D Compress(Texture2D src)
+        {
+            // Compressメソッドで圧縮する準備として、幅・高さを4の倍数にする必要があります。
+            // 最も近い4の倍数を求めます。
+            var widthX4 = (src.width + 2) / 4 * 4;
+            var heightX4 = (src.height + 2) / 4 * 4;
+            widthX4 = Math.Max(4, widthX4); // 幅・高さが 1ピクセルのケースで 0 に丸められてしまうのを防止します。
+            heightX4 = Math.Max(4, heightX4);
+
+            // テクスチャをリサイズします。
+            // 参考: https://light11.hatenadiary.com/entry/2018/04/19/194015
+            var rt = RenderTexture.GetTemporary(widthX4, heightX4);
+            Graphics.Blit(src, rt);
+            var prevRt = RenderTexture.active;
+            RenderTexture.active = rt;
+            var dst = new Texture2D(widthX4, heightX4);
+            dst.ReadPixels(new Rect(0, 0, widthX4, heightX4), 0, 0);
+            dst.Apply();
+            RenderTexture.active = prevRt;
+            RenderTexture.ReleaseTemporary(rt);
+
+            // 圧縮のキモです。
+            dst.Compress(true);
+            return dst;
         }
     }
 }
